Rename test event classes and enhance documentation
Renamed the class `Pippo` to `TestEventForAsync` in the
`NEventStore.Persistence.AcceptanceTests.Async` namespace,
and `Pippo` to `TestEvent` in the
`NEventStore.Persistence.AcceptanceTests` namespace,
updating all relevant instances in the test code for clarity.

Added XML documentation comments in the `IAsyncObserver<T>`
interface to clarify the `cancellationToken` parameter.

Enhanced documentation in the `ICommitEvents` and
`ICommitEventsAsync` interfaces to explain the nullable
`ICommit` return type.

Included remarks in the `Commit` and `CommitAsync` methods
of the `OptimisticEventStore` class to describe the role of
pipeline hooks in the commit process.